refactor: изменены названия импортов и модуля
