+++ app/keyboards/admin/common_kb.py
@@ -10,7 +10,7 @@ def admin_panel_kb() -> InlineKeyboardMarkup:
             callback_data='admin rent',
         )
     )
-    kb.add(InlineKeyboardButton(text='Блок Картины', callback_data='admin picture'))
+    kb.add(InlineKeyboardButton(text='Блок Картины', callback_data='admin pictures'))
     kb.add(
         InlineKeyboardButton(text='Блок О Проекте', callback_data='admin about project')
     )
